改进体验

+++ src/views/cloud-code/index.vue
@@ -2,7 +2,7 @@
  * @Author: Aardpro
  * @Date: 2021-04-11 10:38:30
  * @LastEditors: Aardpro
- * @LastEditTime: 2021-04-12 22:31:49
+ * @LastEditTime: 2021-04-17 12:00:09
  * @Description: 云端代码
 -->
 <template>
@@ -13,6 +13,7 @@
       :key="index"
       class="row code-list"
       :class="{ title: item.isTitle }"
+      @click="askEdit(item)"
     >
       <div class="col-1">
         <span v-if="!item.isTitle" v-text="index"></span><span v-else>#</span>
@@ -22,10 +23,18 @@
       <div class="col-2">{{ item.createtime }}</div>
       <div class="col-2">{{ item.modifytime }}</div>
       <div class="col-2 flex-right">
-        <template v-if="!item.isTitle"
-          ><a @click="askDelete(item)">删除</a>
-          <a @click="askEdit(item)">编辑</a></template
-        ><span v-else>操作</span>
+        <span v-if="item.isTitle">操作</span>
+        <template v-else>
+          <a v-if="userRank > 7" @click="(e)=>{e.stopPropagation();switchFeatured(item)}">{{
+            item.featured ? "unFeature me" : "Featur me"
+          }}</a>
+          <a
+            v-if="userRank > 7 || userId === item.user_id"
+            @click="(e)=>{e.stopPropagation();askDelete(item)}"
+            >删除</a
+          >
+          <a @click="(e)=>{e.stopPropagation();askEdit(item);}">运行</a>
+        </template>
       </div>
     </div>
   </template>
@@ -37,13 +46,17 @@
 
 <script>
 import { defineComponent, ref, onBeforeMount } from "vue";
-import { getCodeMylist, deleteCode } from "../../api";
+import { getCodeMylist, deleteCode, saveCode } from "../../api";
 import { safePush } from "../../router";
+import { getUserId, getUserRank } from "../../utils";
+
 export default defineComponent({
   name: "CloudCode",
   components: {},
   props: {},
   setup(props, context) {
+    const userId = ref(getUserId());
+    const userRank = ref(getUserRank());
     const records = ref([
       {
         route: "",
@@ -52,6 +65,8 @@ export default defineComponent({
         route: "功能类型",
         name: "名称",
         isTitle: true,
+        user_id: "",
+        featured: 0,
       },
     ]);
     onBeforeMount(() => {
@@ -66,6 +81,10 @@ export default defineComponent({
         });
     });
     const askDelete = async (item) => {
+      const result = window.confirm("确定删除一条记录吗？");
+      if (!result) {
+        return;
+      }
       const res = await deleteCode(item.id);
       if (!res.code) {
         records.value.splice(
@@ -74,7 +93,14 @@ export default defineComponent({
         );
       }
     };
+    const switchFeatured = async (item) => {
+      item.featured = 1 - item.featured;
+      await saveCode(item.id, { featured: item.featured });
+    };
     const askEdit = async (item) => {
+      if (!item || !item.id) {
+        return;
+      }
       safePush({
         name: item.route,
         params: {
@@ -82,11 +108,14 @@ export default defineComponent({
         },
       });
     };
-    sessionStorage.setItem("FROM","CLOUD-CODE")
+    sessionStorage.setItem("FROM", "CLOUD-CODE");
     return {
       records,
       askDelete,
       askEdit,
+      switchFeatured,
+      userId,
+      userRank,
     };
   },
 });
